Language.Rzk.Syntax.Layout

Documentation

data LayoutDelimiters

Constructors

LayoutDelimiters

Fields

delimSep :: TokSymbol

delimOpen :: Maybe TokSymbol

Nothing for toplevel layout.

delimClose :: Maybe TokSymbol

Nothing for toplevel layout.

layoutWords :: [(TokSymbol, LayoutDelimiters)]

layoutStopWords :: [TokSymbol]

layoutOpen :: [TokSymbol]

layoutClose :: [TokSymbol]

layoutSep :: [TokSymbol]

parenOpen :: [TokSymbol]

parenClose :: [TokSymbol]

layoutError

Arguments

:: [Token]
Remaining tokens.

-> String
Error message.

-> a

Report an error during layout resolution.

resolveLayout

Arguments

:: Bool
Whether to use top-level layout.

-> [Token]
Token stream before layout resolution.

-> [Token]
Token stream after layout resolution.

Replace layout syntax with explicit layout tokens.

type Position = Posn

type Line = Int

type Column = Int

data Block

Entry of the layout stack.

Constructors

Implicit LayoutDelimiters Status Column
An implicit layout block with its start column.

Explicit

indentation :: Block -> Column

Get current indentation. 0 if we are in an explicit block.

isImplicit :: Block -> Bool

Check if s block is implicit.

data Status

Constructors

Tentative
A layout column that has not been confirmed by a line break

Definitive
A layout column that has been confirmed by a line break.

addImplicit

Arguments

:: LayoutDelimiters
Delimiters of the new block.

-> Position
Position of the layout keyword.

-> Position
Position of the token following the layout keword.

-> [Block]

-> [Block]

Add a new implicit layout block.

confirm :: Column -> [Block] -> [Block]

Confirm tentative blocks that are not more indented than col.

afterPrev :: Maybe Token -> Position

Get the position immediately to the right of the given token. If no token is given, gets the first position in the file.

nextPos :: Token -> Position

Get the position immediately to the right of the given token.

tokenLength :: Token -> Int

Get the number of characters in the token.

sToken :: Position -> TokSymbol -> Token

Create a position symbol token.

line :: Token -> Line

Get the line number of a token.

column :: Token -> Column

Get the column number of a token.

newLine :: Maybe Token -> Token -> Bool

Is the following token on a new line?

isLayout :: Token -> Maybe LayoutDelimiters

Check if a word is a layout start token.

isTokenIn :: [TokSymbol] -> Token -> Bool

Check if a token is one of the given symbols.

isStop :: Token -> Bool

Check if a token is a layout stop token.

isLayoutOpen :: Token -> Bool

Check if a token is the layout open token.

isLayoutSep :: Token -> Bool

Check if a token is the layout separator token.

isLayoutClose :: Token -> Bool

Check if a token is the layout close token.

isParenOpen :: Token -> Bool

Check if a token is an opening parenthesis.

isParenClose :: Token -> Bool

Check if a token is a closing parenthesis.