Language.Rzk.Syntax.Lex

Documentation

alex_tab_size :: Int

alex_base :: AlexAddr

alex_table :: AlexAddr

alex_check :: AlexAddr

alex_deflt :: AlexAddr

alex_accept :: Array Int (AlexAcc user)

alex_actions :: Array Int (Posn -> String -> Token)

alex_action_3 :: Posn -> String -> Token

alex_action_4 :: Posn -> String -> Token

alex_action_5 :: Posn -> String -> Token

alex_action_6 :: Posn -> String -> Token

alex_action_7 :: Posn -> String -> Token

data AlexAddr

Constructors

AlexA# Addr#

alexIndexInt16OffAddr :: AlexAddr -> Int# -> Int#

alexIndexInt32OffAddr :: AlexAddr -> Int# -> Int#

quickIndex :: Array Int (AlexAcc (Any :: Type)) -> Int -> AlexAcc (Any :: Type)

data AlexReturn a

Constructors

AlexEOF

AlexError !AlexInput

AlexSkip !AlexInput !Int

AlexToken !AlexInput !Int a

alexScan :: (Posn, Char, [Byte], String) -> Int -> AlexReturn (Posn -> String -> Token)

alexScanUser :: t -> (Posn, Char, [Byte], String) -> Int -> AlexReturn (Posn -> String -> Token)

alex_scan_tkn :: t1 -> t2 -> Int# -> AlexInput -> Int# -> AlexLastAcc -> (AlexLastAcc, AlexInput)

data AlexLastAcc

Constructors

AlexNone

AlexLastAcc !Int !AlexInput !Int

AlexLastSkip !AlexInput !Int

data AlexAcc user

Constructors

AlexAccNone

AlexAcc Int

AlexAccSkip

tok :: (String -> Tok) -> Posn -> String -> Token

Create a token with position.

data Tok

Token without position.

Constructors

TK !TokSymbol
Reserved word or symbol.

TL !String
String literal.

TI !String
Integer literal.

TV !String
Identifier.

TD !String
Float literal.

TC !String
Character literal.

T_VarIdentToken !String

T_HoleIdentToken !String

pattern TS :: String -> Int -> Tok

Smart constructor for Tok for the sake of backwards compatibility.

data TokSymbol

Keyword or symbol tokens have a unique ID.

Constructors

TokSymbol

Fields

tsText :: String

Keyword or symbol text.

tsID :: !Int

Unique ID.

data Token

Token with position.

Constructors

PT Posn Tok

Err Posn

printPosn :: Posn -> String

Pretty print a position.

tokenPos :: [Token] -> String

Pretty print the position of the first token in the list.

tokenPosn :: Token -> Posn

Get the position of a token.

tokenLineCol :: Token -> (Int, Int)

Get line and column of a token.

posLineCol :: Posn -> (Int, Int)

Get line and column of a position.

mkPosToken :: Token -> ((Int, Int), String)

Convert a token into "position token" form.

tokenText :: Token -> String

Convert a token to its text.

prToken :: Token -> String

Convert a token to a string.

data BTree

Finite map from text to token organized as binary search tree.

Constructors

N
Nil (leaf).

B String Tok BTree BTree
Binary node.

eitherResIdent :: (String -> Tok) -> String -> Tok

Convert potential keyword into token or use fallback conversion.

resWords :: BTree

The keywords and symbols of the language organized as binary search tree.

unescapeInitTail :: String -> String

Unquote string literal.

data Posn

Constructors

Pn !Int !Int !Int

alexStartPos :: Posn

alexMove :: Posn -> Char -> Posn

type Byte = Word8

type AlexInput = (Posn, Char, [Byte], String)

tokens :: String -> [Token]

alexGetByte :: AlexInput -> Maybe (Byte, AlexInput)

alexInputPrevChar :: AlexInput -> Char

utf8Encode :: Char -> [Word8]

Encode a Haskell String to a list of Word8 values, in UTF8 format.