Rzk.Format

Description

The formatter is designed in a way that can be consumed both by the CLI and the LSP server.

Documentation

data FormattingEdit

All indices are 1-based (as received from the lexer) Note: LSP uses 0-based indices

Constructors

FormattingEdit Int Int Int Int String

formatTextEdits :: String -> [FormattingEdit]

format :: String -> String

Format Rzk code, returning the formatted version.

formatFile :: FilePath -> IO String

Format Rzk code from a file

formatFileWrite :: FilePath -> IO ()

Format the file and write the result back to the file.

isWellFormatted :: String -> Bool

Check if the given Rzk source code is well formatted. This is useful for automation tasks.

isWellFormattedFile :: FilePath -> IO Bool

Same as isWellFormatted, but reads the source code from a file.