Rzk.Main

Documentation

parseStdin :: IO Module

globNonEmpty :: FilePath -> IO [FilePath]

Finds matches to the given pattern in the current working directory. **NOTE:** throws exception when glob returns an empty list.

extractFilesFromRzkYaml :: FilePath -> IO [FilePath]

expandRzkPathsOrYaml :: [FilePath] -> IO [FilePath]

Given a list of file paths (possibly including globs), expands the globs (if any) or tries to read the list of files from the rzk.yaml file (if no paths are given). Glob patterns in rzk.yaml are also expanded.

parseRzkFilesOrStdin :: [FilePath] -> IO [(FilePath, Module)]

typecheckString :: String -> Either String String