Rzk.TypeCheck

Documentation

>>> :set -XOverloadedStrings

defaultTypeCheck :: TypeCheck var a -> Either (TypeErrorInScopedContext var) a

data Decl var

Constructors

Decl

Fields

declName :: var

declType :: TermT var

declValue :: Maybe (TermT var)

declIsAssumption :: Bool

declUsedVars :: [var]

declLocation :: Maybe LocationInfo

type Decl' = Decl VarIdent

typecheckModulesWithLocationIncremental

Arguments

:: [(FilePath, [Decl'])]
Cached declarations (only those that do not need rechecking).

-> [(FilePath, Module)]
New modules to check

-> TypeCheck VarIdent ([(FilePath, [Decl'])], [TypeErrorInScopedContext VarIdent])

typecheckModulesWithLocation' :: [(FilePath, Module)] -> TypeCheck VarIdent ([(FilePath, [Decl'])], [TypeErrorInScopedContext VarIdent])

typecheckModulesWithLocation :: [(FilePath, Module)] -> TypeCheck VarIdent [(FilePath, [Decl'])]

typecheckModules :: [Module] -> TypeCheck VarIdent [Decl']

typecheckModuleWithLocation :: (FilePath, Module) -> TypeCheck VarIdent ([Decl'], [TypeErrorInScopedContext VarIdent])

countCommands :: Integral a => [Command] -> a

typecheckModule :: Maybe FilePath -> Module -> TypeCheck VarIdent ([Decl'], [TypeErrorInScopedContext VarIdent])

splitSectionCommands :: SectionName -> [Command] -> TypeCheck var ([Command], [Command])

setOption :: String -> String -> TypeCheck var a -> TypeCheck var a

unsetOption :: String -> TypeCheck var a -> TypeCheck var a

paramToParamDecl :: Param -> TypeCheck var [ParamDecl]

addParamDecls :: [ParamDecl] -> Term -> Term

addParams :: [Param] -> Term -> Term

data TypeError var

Constructors

TypeErrorOther String

TypeErrorUnify (TermT var) (TermT var) (TermT var)

TypeErrorUnifyTerms (TermT var) (TermT var)

TypeErrorNotPair (TermT var) (TermT var)

TypeErrorNotFunction (TermT var) (TermT var)

TypeErrorUnexpectedLambda (Term var) (TermT var)

TypeErrorUnexpectedPair (Term var) (TermT var)

TypeErrorUnexpectedRefl (Term var) (TermT var)

TypeErrorCannotInferBareLambda (Term var)

TypeErrorCannotInferBareRefl (Term var)

TypeErrorUndefined var

TypeErrorTopeNotSatisfied [TermT var] (TermT var)

TypeErrorTopesNotEquivalent (TermT var) (TermT var)

TypeErrorInvalidArgumentType (Term var) (TermT var)

TypeErrorDuplicateTopLevel [VarIdent] VarIdent

TypeErrorUnusedVariable var (TermT var)

TypeErrorUnusedUsedVariables [var] var

TypeErrorImplicitAssumption (var, TermT var) var

data TypeErrorInContext var

Constructors

TypeErrorInContext

Fields

typeErrorError :: TypeError var

typeErrorContext :: Context var

data TypeErrorInScopedContext var

Constructors

PlainTypeError (TypeErrorInContext var)

ScopedTypeError (Maybe VarIdent) (TypeErrorInScopedContext (Inc var))

type TypeError' = TypeError VarIdent

ppTypeError' :: TypeError' -> String

ppTypeErrorInContext :: OutputDirection -> TypeErrorInContext VarIdent -> String

ppTypeErrorInScopedContextWith' :: OutputDirection -> [VarIdent] -> [VarIdent] -> TypeErrorInScopedContext VarIdent -> String

ppTypeErrorInScopedContext' :: OutputDirection -> TypeErrorInScopedContext VarIdent -> String

issueWarning :: String -> TypeCheck var ()

fromTypeError :: TypeError var -> TypeCheck var (TypeErrorInScopedContext var)

issueTypeError :: TypeError var -> TypeCheck var a

panicImpossible :: String -> a

data Action var

Constructors

ActionTypeCheck (Term var) (TermT var)

ActionUnify (TermT var) (TermT var) (TermT var)

ActionUnifyTerms (TermT var) (TermT var)

ActionInfer (Term var)

ActionContextEntailedBy [TermT var] (TermT var)

ActionContextEntails [TermT var] (TermT var)

ActionContextEquiv [TermT var] [TermT var]

ActionWHNF (TermT var)

ActionNF (TermT var)

ActionCheckCoherence (TermT var, TermT var) (TermT var, TermT var)

ActionCloseSection (Maybe SectionName)

type Action' = Action VarIdent

ppTermInContext :: Eq var => TermT var -> TypeCheck var String

ppSomeAction :: Eq var => [(var, Maybe VarIdent)] -> Int -> Action var -> String

ppAction :: Int -> Action' -> String

traceAction' :: Int -> Action' -> a -> a

unsafeTraceAction' :: Int -> Action var -> a -> a

data LocationInfo

Constructors

LocationInfo

Fields

locationFilePath :: Maybe FilePath

locationLine :: Maybe Int

data Verbosity

Constructors

Debug

Normal

Silent

trace' :: Verbosity -> Verbosity -> String -> a -> a

traceTypeCheck :: Verbosity -> String -> TypeCheck var a -> TypeCheck var a

localVerbosity :: Verbosity -> TypeCheck var a -> TypeCheck var a

localRenderBackend :: Maybe RenderBackend -> TypeCheck var a -> TypeCheck var a

data Covariance

Constructors

Covariant
Positive position.

Contravariant
Negative position

data RenderBackend

Constructors

RenderSVG

RenderLaTeX

data ScopeInfo var

Constructors

ScopeInfo

Fields

scopeName :: Maybe SectionName

scopeVars :: [(var, VarInfo var)]

addVarToScope :: var -> VarInfo var -> ScopeInfo var -> ScopeInfo var

data VarInfo var

Constructors

VarInfo

Fields

varType :: TermT var

varValue :: Maybe (TermT var)

varOrig :: Maybe VarIdent

varIsAssumption :: Bool

varDeclaredAssumptions :: [var]

varLocation :: Maybe LocationInfo

data Context var

Constructors

Context

Fields

localScopes :: [ScopeInfo var]

localTopes :: [TermT var]

localTopesNF :: [TermT var]

localTopesNFUnion :: [[TermT var]]

localTopesEntailBottom :: Bool

actionStack :: [Action var]

currentCommand :: Maybe Command

location :: Maybe LocationInfo

verbosity :: Verbosity

covariance :: Covariance

renderBackend :: Maybe RenderBackend

addVarInCurrentScope :: var -> VarInfo var -> Context var -> Context var

emptyContext :: Context var

askCurrentScope :: TypeCheck var (ScopeInfo var)

varInfos :: Context var -> [(var, VarInfo var)]

varTypes :: Context var -> [(var, TermT var)]

varValues :: Context var -> [(var, Maybe (TermT var))]

varOrigs :: Context var -> [(var, Maybe VarIdent)]

withPartialDecls :: TypeCheck VarIdent ([Decl'], [err]) -> TypeCheck VarIdent ([Decl'], [err]) -> TypeCheck VarIdent ([Decl'], [err])

withSection :: Maybe SectionName -> TypeCheck VarIdent ([Decl VarIdent], [TypeErrorInScopedContext VarIdent]) -> TypeCheck VarIdent ([Decl VarIdent], [TypeErrorInScopedContext VarIdent]) -> TypeCheck VarIdent ([Decl VarIdent], [TypeErrorInScopedContext VarIdent])

startSection :: Maybe SectionName -> TypeCheck VarIdent a -> TypeCheck VarIdent a

endSection :: [TypeErrorInScopedContext VarIdent] -> TypeCheck VarIdent ([Decl'], [TypeErrorInScopedContext VarIdent])

scopeToDecls :: Eq var => [TypeErrorInScopedContext var] -> ScopeInfo var -> TypeCheck var ([Decl var], [TypeErrorInScopedContext var])

insertExplicitAssumptionFor :: Eq var => var -> (var, VarInfo var) -> TermT var -> TermT var

insertExplicitAssumptionFor' :: Eq var => var -> (var, VarInfo var) -> VarInfo var -> VarInfo var

makeAssumptionExplicit :: Eq var => (var, VarInfo var) -> [(var, VarInfo var)] -> TypeCheck var (Bool, [(var, VarInfo var)])

collectScopeDecls :: Eq var => [TypeErrorInScopedContext var] -> [(var, VarInfo var)] -> [(var, VarInfo var)] -> TypeCheck var ([Decl var], [TypeErrorInScopedContext var])

abstractAssumption :: Eq var => (var, VarInfo var) -> Decl var -> Decl var

data OutputDirection

Constructors

TopDown

BottomUp

block :: OutputDirection -> [String] -> String

namedBlock :: OutputDirection -> String -> [String] -> String

ppContext' :: OutputDirection -> Context VarIdent -> String

doesShadowName :: VarIdent -> TypeCheck var [VarIdent]

checkTopLevelDuplicate :: VarIdent -> TypeCheck var ()

checkNameShadowing :: VarIdent -> TypeCheck var ()

withLocation :: LocationInfo -> TypeCheck var a -> TypeCheck var a

withCommand :: Command -> TypeCheck VarIdent ([Decl'], [TypeErrorInScopedContext VarIdent]) -> TypeCheck VarIdent ([Decl'], [TypeErrorInScopedContext VarIdent])

localDecls :: [Decl VarIdent] -> TypeCheck VarIdent a -> TypeCheck VarIdent a

localDeclsPrepared :: [Decl VarIdent] -> TypeCheck VarIdent a -> TypeCheck VarIdent a

localDecl :: Decl VarIdent -> TypeCheck VarIdent a -> TypeCheck VarIdent a

localDeclPrepared :: Decl VarIdent -> TypeCheck VarIdent a -> TypeCheck VarIdent a

type TypeCheck var = ReaderT (Context var) (Except (TypeErrorInScopedContext var))

freeVarsT_ :: Eq var => TermT var -> TypeCheck var [var]

traceStartAndFinish :: Show a => String -> a -> a

entail :: Eq var => [TermT var] -> TermT var -> Bool

entailM :: Eq var => [TermT var] -> TermT var -> TypeCheck var Bool

entailTraceM :: Eq var => [TermT var] -> TermT var -> TypeCheck var Bool

nubTermT :: Eq var => [TermT var] -> [TermT var]

saturateTopes :: Eq var => [TermT var] -> [TermT var] -> [TermT var]

saturateWith :: (a -> [a] -> Bool) -> ([a] -> [a] -> [a]) -> [a] -> [a]

generateTopes :: Eq var => [TermT var] -> [TermT var] -> [TermT var]

generateTopesForPoints :: Eq var => [TermT var] -> [TermT var]

generateTopesForPointsM :: Eq var => [TermT var] -> TypeCheck var [TermT var]

allTopePoints :: Eq var => TermT var -> [TermT var]

topePoints :: TermT var -> [TermT var]

subPoints :: TermT var -> [TermT var]

simplifyLHSwithDisjunctions :: Eq var => [TermT var] -> [[TermT var]]

Simplify the context, including disjunctions. See also simplifyLHS.

simplifyLHS :: Eq var => [TermT var] -> [TermT var]

Simplify the context, except disjunctions. See also simplifyLHSwithDisjunctions.

solveRHSM :: Eq var => [TermT var] -> TermT var -> TypeCheck var Bool

solveRHS :: Eq var => [TermT var] -> TermT var -> Bool

checkTope :: Eq var => TermT var -> TypeCheck var Bool

checkTopeEntails :: Eq var => TermT var -> TypeCheck var Bool

checkEntails :: Eq var => TermT var -> TermT var -> TypeCheck var Bool

contextEntailedBy :: Eq var => TermT var -> TypeCheck var ()

contextEntails :: Eq var => TermT var -> TypeCheck var ()

topesEquiv :: Eq var => TermT var -> TermT var -> TypeCheck var Bool

contextEquiv :: Eq var => [TermT var] -> TypeCheck var ()

switchVariance :: TypeCheck var a -> TypeCheck var a

enterScopeContext :: Maybe VarIdent -> TermT var -> Context var -> Context (Inc var)

enterScope :: Maybe VarIdent -> TermT var -> TypeCheck (Inc var) b -> TypeCheck var b

performing :: Eq var => Action var -> TypeCheck var a -> TypeCheck var a

stripTypeRestrictions :: TermT var -> TermT var

etaMatch :: Eq var => Maybe (TermT var) -> TermT var -> TermT var -> TypeCheck var (TermT var, TermT var)

Perform at most one \(\eta\)-expansion at the top-level to assist unification.

etaExpand :: Eq var => TermT var -> TypeCheck var (TermT var)

inCubeLayer :: Eq var => TermT var -> TypeCheck var Bool

inTopeLayer :: Eq var => TermT var -> TypeCheck var Bool

tryRestriction :: Eq var => TermT var -> TypeCheck var (Maybe (TermT var))

whnfT :: Eq var => TermT var -> TypeCheck var (TermT var)

Compute a typed term to its WHNF.

>>> unsafeTypeCheck' $ whnfT "(\\ (x : Unit) -> x) unit"
unit : Unit

nfTope :: Eq var => TermT var -> TypeCheck var (TermT var)

nfT :: Eq var => TermT var -> TypeCheck var (TermT var)

Compute a typed term to its NF.

>>> unsafeTypeCheck' $ nfT "(\\ (x : Unit) -> x) unit"
unit : Unit

checkDefinedVar :: VarIdent -> TypeCheck VarIdent ()

valueOfVar :: Eq var => var -> TypeCheck var (Maybe (TermT var))

typeOfVar :: Eq var => var -> TypeCheck var (TermT var)

typeOfUncomputed :: Eq var => TermT var -> TypeCheck var (TermT var)

typeOf :: Eq var => TermT var -> TypeCheck var (TermT var)

unifyTopes :: Eq var => TermT var -> TermT var -> TypeCheck var ()

inAllSubContexts :: TypeCheck var () -> TypeCheck var () -> TypeCheck var ()

unify :: Eq var => Maybe (TermT var) -> TermT var -> TermT var -> TypeCheck var ()

unifyViaDecompose :: Eq var => TermT var -> TermT var -> TypeCheck var ()

unifyInCurrentContext :: Eq var => Maybe (TermT var) -> TermT var -> TermT var -> TypeCheck var ()

unifyTypes :: Eq var => TermT var -> TermT var -> TermT var -> TypeCheck var ()

unifyTerms :: Eq var => TermT var -> TermT var -> TypeCheck var ()

localTope :: Eq var => TermT var -> TypeCheck var a -> TypeCheck var a

universeT :: TermT var

cubeT :: TermT var

topeT :: TermT var

topeEQT :: TermT var -> TermT var -> TermT var

topeLEQT :: TermT var -> TermT var -> TermT var

topeOrT :: TermT var -> TermT var -> TermT var

topeAndT :: TermT var -> TermT var -> TermT var

cubeProductT :: TermT var -> TermT var -> TermT var

cubeUnitT :: TermT var

cubeUnitStarT :: TermT var

typeUnitT :: TermT var

unitT :: TermT var

cube2T :: TermT var

cube2_0T :: TermT var

cube2_1T :: TermT var

topeTopT :: TermT var

topeBottomT :: TermT var

recBottomT :: TermT var

typeRestrictedT :: TermT var -> [(TermT var, TermT var)] -> TermT var

lambdaT :: TermT var -> Maybe VarIdent -> Maybe (TermT var, Maybe (Scope TermT var)) -> Scope TermT var -> TermT var

appT :: TermT var -> TermT var -> TermT var -> TermT var

pairT :: TermT var -> TermT var -> TermT var -> TermT var

firstT :: TermT var -> TermT var -> TermT var

secondT :: TermT var -> TermT var -> TermT var

reflT :: TermT var -> Maybe (TermT var, Maybe (TermT var)) -> TermT var

typeFunT :: Maybe VarIdent -> TermT var -> Maybe (Scope TermT var) -> Scope TermT var -> TermT var

typeSigmaT :: Maybe VarIdent -> TermT var -> Scope TermT var -> TermT var

recOrT :: TermT var -> [(TermT var, TermT var)] -> TermT var

typeIdT :: TermT var -> Maybe (TermT var) -> TermT var -> TermT var

idJT :: TermT var -> TermT var -> TermT var -> TermT var -> TermT var -> TermT var -> TermT var -> TermT var

typeAscT :: TermT var -> TermT var -> TermT var

typecheck :: Eq var => Term var -> TermT var -> TypeCheck var (TermT var)

inferAs :: Eq var => TermT var -> Term var -> TypeCheck var (TermT var)

infer :: Eq var => Term var -> TypeCheck var (TermT var)

checkCoherence :: Eq var => (TermT var, TermT var) -> (TermT var, TermT var) -> TypeCheck var ()

inferStandalone :: Eq var => Term var -> Either (TypeErrorInScopedContext var) (TermT var)

unsafeInferStandalone' :: Term' -> TermT'

unsafeTypeCheck' :: TypeCheck VarIdent a -> a

type PointId = String

type ShapeId = [PointId]

cube2powerT :: Int -> TermT var

splits :: [a] -> [([a], [a])]

verticesFrom :: [TermT var] -> [(ShapeId, TermT var)]

subTopes2 :: Int -> TermT var -> [(ShapeId, TermT var)]

cubeSubTopes :: [(ShapeId, TermT (Inc var))]

limitLength :: Int -> String -> String

renderObjectsFor :: Eq var => String -> Int -> TermT var -> TermT var -> TypeCheck var [(ShapeId, RenderObjectData)]

componentWiseEQT :: Int -> TermT var -> TermT var -> TermT var

renderObjectsInSubShapeFor :: Eq var => String -> Int -> [var] -> var -> TermT var -> TermT var -> TermT var -> TypeCheck var [(ShapeId, RenderObjectData)]

renderForSubShapeSVG :: Eq var => String -> Int -> [var] -> var -> TermT var -> TermT var -> TermT var -> TypeCheck var String

renderForSVG :: Eq var => String -> Int -> TermT var -> TermT var -> TypeCheck var String

renderTermSVGFor

Arguments

:: Eq var

=> String
Main color.

-> Int
Accumulated dimensions so far (from 0 to 3).

-> (Maybe (TermT var, TermT var), [var])
Accumulated point term (and its time).

-> TermT var
Term to render.

-> TypeCheck var (Maybe String)

renderTermSVG :: Eq var => TermT var -> TypeCheck var (Maybe String)

renderTermSVG' :: Eq var => TermT var -> TypeCheck var (Maybe String)

type Point2D a = (a, a)

type Point3D a = (a, a, a)

type Edge3D a = (Point3D a, Point3D a)

type Face3D a = (Point3D a, Point3D a, Point3D a)

type Volume3D a = (Point3D a, Point3D a, Point3D a, Point3D a)

data CubeCoords2D a b

Constructors

CubeCoords2D

Fields

vertices :: [(Point3D a, Point2D b)]

edges :: [(Edge3D a, (Point2D b, Point2D b))]

faces :: [(Face3D a, (Point2D b, Point2D b, Point2D b))]

volumes :: [(Volume3D a, (Point2D b, Point2D b, Point2D b, Point2D b))]

data Matrix3D a

Constructors

Matrix3D a a a a a a a a a

data Matrix4D a

Constructors

Matrix4D a a a a a a a a a a a a a a a a

data Vector3D a

Constructors

Vector3D a a a

data Vector4D a

Constructors

Vector4D a a a a

rotateX :: Floating a => a -> Matrix3D a

rotateY :: Floating a => a -> Matrix3D a

rotateZ :: Floating a => a -> Matrix3D a

data Camera a

Constructors

Camera

Fields

cameraPos :: Point3D a

cameraFoV :: a

cameraAspectRatio :: a

cameraAngleY :: a

cameraAngleX :: a

viewRotateX :: Floating a => Camera a -> Matrix4D a

viewRotateY :: Floating a => Camera a -> Matrix4D a

viewTranslate :: Num a => Camera a -> Matrix4D a

project2D :: Floating a => Camera a -> Matrix4D a

matrixVectorMult4D :: Num a => Matrix4D a -> Vector4D a -> Vector4D a

matrix3Dto4D :: Num a => Matrix3D a -> Matrix4D a

fromAffine :: Fractional a => Vector4D a -> (Point2D a, a)

point3Dto2D :: Floating a => Camera a -> a -> Point3D a -> (Point2D a, a)

data RenderObjectData

Constructors

RenderObjectData

Fields

renderObjectDataLabel :: String

renderObjectDataFullLabel :: String

renderObjectDataColor :: String

renderCube :: (Floating a, Show a) => Camera a -> a -> (String -> Maybe RenderObjectData) -> String

defaultCamera :: Floating a => Camera a

Orphan instances

IsString TermT'
Parse and unsafeInferStandalone'.

Instance details

Methods

fromString :: String -> TermT'